fix: fix registry module caching using normalized registry source

+++ internal/hcl/modules/loader_test.go
@@ -9,14 +9,18 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func testLoaderExpectedModules(t *testing.T, path string, expectedModules []*ManifestModule) {
-	err := os.RemoveAll(filepath.Join(path, ".infracost"))
-	assert.NoError(t, err)
+func testLoaderE2E(t *testing.T, path string, expectedModules []*ManifestModule, cleanup bool) {
+	if cleanup {
+		err := os.RemoveAll(filepath.Join(path, ".infracost"))
+		assert.NoError(t, err)
+	}
 
 	moduleLoader := NewModuleLoader(path)
 
 	manifest, err := moduleLoader.Load()
-	assert.NoError(t, err)
+	if !assert.NoError(t, err) {
+		t.FailNow()
+	}
 
 	sort.Slice(expectedModules, func(i, j int) bool {
 		return expectedModules[i].Key < expectedModules[j].Key
@@ -36,7 +40,7 @@ func TestNestedModules(t *testing.T) {
 		t.Skip("Skipping test in short mode")
 	}
 
-	testLoaderExpectedModules(t, "./testdata/nested_modules", []*ManifestModule{
+	testLoaderE2E(t, "./testdata/nested_modules", []*ManifestModule{
 		{
 			Key:    "local-module",
 			Source: "./modules/local-module",
@@ -69,7 +73,7 @@ func TestNestedModules(t *testing.T) {
 			Source: "git::https://github.com/terraform-aws-modules/terraform-aws-sns.git",
 			Dir:    ".infracost/terraform_modules/local-module.nested-git-module",
 		},
-	})
+	}, true)
 }
 
 func TestSubmodules(t *testing.T) {
@@ -77,7 +81,7 @@ func TestSubmodules(t *testing.T) {
 		t.Skip("Skipping test in short mode")
 	}
 
-	testLoaderExpectedModules(t, "./testdata/submodules", []*ManifestModule{
+	testLoaderE2E(t, "./testdata/submodules", []*ManifestModule{
 		{
 			Key:     "registry-submodule",
 			Source:  "registry.terraform.io/terraform-aws-modules/route53/aws//modules/zones",
@@ -89,7 +93,7 @@ func TestSubmodules(t *testing.T) {
 			Source: "git::https://github.com/terraform-aws-modules/terraform-aws-route53.git//modules/zones",
 			Dir:    ".infracost/terraform_modules/git-submodule/modules/zones",
 		},
-	})
+	}, true)
 }
 
 func TestModuleMultipleUses(t *testing.T) {
@@ -97,7 +101,7 @@ func TestModuleMultipleUses(t *testing.T) {
 		t.Skip("Skipping test in short mode")
 	}
 
-	testLoaderExpectedModules(t, "./testdata/module_multiple_uses", []*ManifestModule{
+	testLoaderE2E(t, "./testdata/module_multiple_uses", []*ManifestModule{
 		{
 			Key:     "registry-module-1",
 			Source:  "registry.terraform.io/terraform-aws-modules/ec2-instance/aws",
@@ -110,5 +114,44 @@ func TestModuleMultipleUses(t *testing.T) {
 			Version: "3.4.0",
 			Dir:     ".infracost/terraform_modules/registry-module-2",
 		},
-	})
+	}, true)
+}
+
+func TestWithCachedModules(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test in short mode")
+	}
+
+	testLoaderE2E(t, "./testdata/with_cached_modules", []*ManifestModule{
+		{
+			Key:    "local-module",
+			Source: "./modules/local-module",
+			Dir:    "modules/local-module",
+		},
+		{
+			Key:     "registry-module",
+			Source:  "registry.terraform.io/terraform-aws-modules/ec2-instance/aws",
+			Version: "3.4.0",
+			Dir:     ".infracost/terraform_modules/registry-module",
+		},
+		{
+			Key:    "git-module",
+			Source: "git::https://github.com/terraform-aws-modules/terraform-aws-ec2-instance.git",
+			Dir:    ".infracost/terraform_modules/git-module",
+		},
+	}, false)
+
+	// Check that the modules were not overwritten
+	regModContents, err := os.ReadFile("./testdata/with_cached_modules/.infracost/terraform_modules/registry-module/main.tf")
+	if !assert.NoError(t, err) {
+		t.FailNow()
+	}
+
+	gitModContents, err := os.ReadFile("./testdata/with_cached_modules/.infracost/terraform_modules/git-module/main.tf")
+	if !assert.NoError(t, err) {
+		t.FailNow()
+	}
+
+	assert.Equal(t, string(regModContents), "// Placeholder file\n")
+	assert.Equal(t, string(gitModContents), "// Placeholder file\n")
 }
